requirements.txt
discord == 2.0.0
python-dotenv == 0.21.0